backed out changes that made SingleHopClientExecutor mockable
