+++ geode-core/src/main/java/org/apache/geode/cache/client/internal/ClientMetadataService.java
@@ -539,7 +539,7 @@ public class ClientMetadataService {
           }
         }
       };
-      SingleHopClientExecutorImpl.getInstance().submitTask(fetchTask);
+      SingleHopClientExecutor.submitTask(fetchTask);
     }
   }
 
@@ -652,7 +652,7 @@ public class ClientMetadataService {
           }
         }
       };
-      SingleHopClientExecutorImpl.getInstance().submitTask(fetchTask);
+      SingleHopClientExecutor.submitTask(fetchTask);
     }
   }
