+++ geode-core/src/main/java/org/apache/geode/cache/client/internal/ExecuteFunctionOp.java
@@ -82,7 +82,7 @@ public class ExecuteFunctionOp {
       List callableTasks = constructAndGetFunctionTasks(pool, function, args, memberMappedArg,
           hasResult, rc, isFnSerializationReqd, attributes);
 
-      SingleHopClientExecutorImpl.getInstance().submitAll(callableTasks);
+      SingleHopClientExecutor.submitAll(callableTasks);
 
     } else {
       AbstractOp reexecOp = null;
@@ -149,7 +149,7 @@ public class ExecuteFunctionOp {
       List callableTasks = constructAndGetFunctionTasks(pool, functionId, args, memberMappedArg,
           hasResult, rc, isFnSerializationReqd, isHA, optimizeForWrite, properties);
 
-      SingleHopClientExecutorImpl.getInstance().submitAll(callableTasks);
+      SingleHopClientExecutor.submitAll(callableTasks);
     } else {
       AbstractOp reexecOp = null;
       boolean reexecuteForServ = false;
