+++ geode-core/src/main/java/org/apache/geode/cache/client/internal/PutAllOp.java
@@ -111,7 +111,7 @@ public class PutAllOp {
         new HashMap<ServerLocation, RuntimeException>();
     PutAllPartialResult result = new PutAllPartialResult(map.size());
     try {
-      Map<ServerLocation, Object> results = SingleHopClientExecutorImpl.getInstance()
+      Map<ServerLocation, Object> results = SingleHopClientExecutor
           .submitBulkOp(callableTasks, cms,
               (LocalRegion) region, failedServers);
       for (Map.Entry<ServerLocation, Object> entry : results.entrySet()) {
